Add Mopidy media playback backend
Plugin talks to Mopidy music player daemon using its WebSocket
interface.

Mopidy controls media playback on a system and coordinates with a
frontend through this plugin.

Mopidy installation must be configured to use the default port #6680.

Currently, only media playback and basic indexing are implemented.
Mopidy is supposed to have `media_dir` set with a path to a folder
with media files which will be automatically added to the tracklist.

Co-authored-by: Florent Revest <frevest@luxoft.com>
Signed-off-by: Oleksandr Kravchuk <oleksandr.kravchuk@pelagicore.com>